refactor(web): move shared helpers into a web::helpers module
Reorganise the web module so that `web::*` modules focus on endpoints and
endpoint-local helpers, while reusable machinery lives under `web::helpers`:

- Move request.rs and csrf.rs into web/helpers/
- Extract the OIDC machinery (discovery/JWKS caching, token validation,
  cookies, claim filtering, safe_return_to, begin_login, exchange_code) into
  web/helpers/oidc.rs, leaving web/oidc.rs with just the callback endpoint and
  the admin_auth middleware
- Update call sites in admin.rs and oauth.rs to the new helper paths
