+++ src/web/mod.rs
@@ -4,10 +4,9 @@ use human_errors::ResultExt;
 use crate::prelude::Services;
 
 mod admin;
-mod csrf;
+mod helpers;
 mod oauth;
 mod oidc;
-mod request;
 mod telemetry;
 mod ui;
 mod webhooks;
